Change DatabaseResolver namespace
